Don't suggest user's to add `:bamboo` in their `:applications`
Elixir 1.4 added application inference which makes
that unecessary. It is generally recommended to not set `:applications`,
but instead only set `:extra_applications`, but since Elixir 1.4 was
released in 2017 I don't think we need to add any instructions for users
still on Elixir 1.3 and earlier.

Additionally if the user only has set some of the applications in
`:applications` there won't be a problem until they try to run a release.

Also adjust the minimum supported version of Elixir to 1.4 to make use
of `:extra_applications`

And change some charlists to strings in the test file.